Finished AStar!!!!!

+++ DirectedWeightedGraph/VertexMapValue.cs
@@ -6,11 +6,11 @@ namespace DirectedWeightedGraph
 {
     class VertexMapValue<T>
     {
-        public int distance { get; set; }
+        public double distance { get; set; }
         public Vertex<T> founder { get; set; }
         public bool wasVisited { get; set; }
-        public int finalDistance { get; set; }
-        public VertexMapValue(int distance, Vertex<T> founder, bool wasVisited, int finalDistance)
+        public double finalDistance { get; set; }
+        public VertexMapValue(double distance, Vertex<T> founder, bool wasVisited, double finalDistance)
         {
             this.distance = distance;
             this.founder = founder;
